Ride Mode
Implemented "Ride mode" basics (Pretty much just shield placement, and debugging transitions between this mode and others).

Adjusted movement and jumping mechanics to be more physics-based.
Holding down the jump button increases jump height/force, rather than just one defined value.

Various other attempts at fine-tuning and bug-fixes.
---
It doesn't seem like a lot (and maybe it isn't), but it took quite a few hours, debugging so much as I went along (including fixing bugs that were already there from past commits), as well as altering the ways some things were already implemented (for instance, the shield is basically always a trigger collider, except when the player steps on it).
May be a small payoff, but I definitely wouldn't like to repeat the amount of time and work it took. Hence, this commit.
Plus, now that these things are implemented, I can (hopefully) just work on fine-tuning them (and their animations), and begin to expand on the levels more.
I've also been thinking about making this more of a platforming game, but keeping the "Runner" scenes for a bit of diversity in gameplay (kind of like "Crash Bandicoot" or "Outland"). The player and shield controls should basically be the same, but without the constant intensity of being hurled at high speeds through the level. May open opportunities to turn this into kind of a puzzle game (like "Nihilumbra" or "Limbo", maybe).